fix(web): address Windows editor PR review
