+++ apps/web/e2e/tests/task/windows-host-embedded-vscode-availability.spec.ts
@@ -1,6 +1,7 @@
 import { test, expect } from "../../fixtures/test-base";
 import type { ApiClient } from "../../helpers/api-client";
 import type { Page } from "@playwright/test";
+import { rewriteBackendHostOS } from "../../helpers/boot-payload";
 import { SessionPage } from "../../pages/session-page";
 
 async function createHostedEditor(apiClient: ApiClient): Promise<void> {
@@ -11,22 +12,6 @@ async function createHostedEditor(apiClient: ApiClient): Promise<void> {
     enabled: true,
   });
   expect(response.ok).toBe(true);
-}
-
-async function rewriteBackendHostOS(page: Page, hostOS: string): Promise<void> {
-  await page.route("**/*", async (route) => {
-    if (route.request().resourceType() !== "document") {
-      await route.continue();
-      return;
-    }
-    const response = await route.fetch();
-    const body = await response.text();
-    const rewritten = body.replace(/"hostOS":"[^"]*"/, `"hostOS":"${hostOS}"`);
-    if (rewritten === body) {
-      throw new Error("The boot document did not contain runtime.hostOS");
-    }
-    await route.fulfill({ response, body: rewritten });
-  });
 }
 
 async function seedTaskAndOpenEditorMenu(
