+++ docs/specs/INDEX.md
@@ -137,7 +137,7 @@ Subscription quota tracking and per-agent cheap-model profile routing.
 | [external-vcs-file-links](ui/external-vcs-file-links.md) | shipped |
 | [task-listing-display-preferences](ui/task-listing-display-preferences.md) | shipped |
 | [task-workspace-content-search](ui/task-workspace-content-search.md) | shipped |
-| [embedded-vscode-windows-availability](ui/embedded-vscode-windows-availability.md) | draft |
+| [embedded-vscode-windows-availability](ui/embedded-vscode-windows-availability.md) | shipped |
 
 ## system-page/ — operational diagnostics & maintenance UI
